% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{squared_exponential_cov}
\alias{squared_exponential_cov}
\alias{squared_exponential_ard_cov}
\alias{linear_cov}
\title{Kernel matrices and their derivatives}
\usage{
squared_exponential_cov(X, Z = NULL, alpha = 1, lengthscale = 1)

squared_exponential_ard_cov(X, Z = NULL, alpha = 1, lengthscale = 1)

linear_cov(X, Z = NULL, bias = 1, deviation = 1)
}
\arguments{
\item{X}{First matrix of observation inputs (N1 x D).}

\item{Z}{Second matrix of observation inputs. By default it is NULL
and treated as equal to X (N2 x D).}

\item{alpha}{Marginal standard deviation hyperparameter. The kernel
is multiplied by the square of this.}

\item{lengthscale}{A vector (D x 1) of lengthscale hyperparameters.
The quadratic terms of the kernel are divided by the squares of these.
If a single number is given, all lengthscales are treated equal
(consider using squared_exponential_cov()).}

\item{bias}{bias hyperparameter.}

\item{deviation}{A vector (D x 1) of hyperparameters describing the deviation of each
input variable, sort of equivalent to regression coefficient.}
}
\value{
Returns a list with the squared exponential kernel matrix and
its first and second derivatives with respect to X (N1 x N2).

Returns a list with the squared exponential kernel matrix and
its first and second derivatives with respect to X (N1 x N2).

Returns a list with the linear kernel matrix and
its first and second derivatives with respect to X (N1 x N2).
}
\description{
A function for computing the squared exponential kernel matrix
and its first and second derivatives with respect to the input data.

A function for computing the ARD squared exponential kernel matrix
and its first and second derivatives with respect to the input data.

A function for computing the linear kernel matrix
and its first and second derivatives with respect to the input data.
}
