% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_diff.R
\name{KL_diff_gaussian}
\alias{KL_diff_gaussian}
\alias{KL_diff2_gaussian}
\alias{KL_diff2_gaussian_linear}
\title{KL-diff and KL-diff2}
\usage{
KL_diff_gaussian(
  y,
  X,
  alpha = 1,
  lengthscale = 1,
  sigma = 1,
  pointwise = FALSE
)

KL_diff2_gaussian(
  y,
  X,
  subset_variables = NULL,
  alpha = 1,
  lengthscale = 1,
  sigma = 1,
  pointwise = FALSE
)

KL_diff2_gaussian_linear(
  y,
  X,
  Xlin,
  subset_variables = NULL,
  alpha = 1,
  lengthscale = 1,
  sigma = 1,
  bias = 1,
  deviation = 1,
  pointwise = FALSE
)
}
\arguments{
\item{y}{Vector (N1) of observation outputs.}

\item{X}{Matrix of observation inputs (N1 x D). Includes the dummy variables.}

\item{alpha}{Marginal standard deviation hyperparameter. The kernel
is multiplied by the square of this.}

\item{lengthscale}{A vector (D x 1) of lengthscale hyperparameters.
The quadratic terms of the kernel are divided by the squares of these.
If a single number is given, all lengthscales are treated equal.}

\item{sigma}{Noise variance hyperparameter.
(NOTE: variance, not standard deviation)}

\item{pointwise}{Logical indicating whether to return KL-diff values
for each observation, or the mean over all observations.}

\item{subset_variables}{Optional vector that represents a subset
of the D variables for which the KL-diff2 values are computed}

\item{Xlin}{Matrix of observation inputs (N1 x D_numeric) excluding
the dummy variables. The linear kernel operates only on these.}

\item{bias}{bias hyperparameter of the linear kernel.}

\item{deviation}{A vector (D x 1) of hyperparameters describing the deviation
of the linear kernel of each
input variable, sort of equivalent to regression coefficient.}
}
\value{
Returns a matrix (N1 x D) or vector (D) of KL-diff values.

Returns an array (N1 x D x D) or matrix (D x D) of KL-diff2 values.

Returns an array (N1 x D x D) or matrix (D x D) of KL-diff2 values.
}
\description{
A function for computing the KL-diff values for all observations
and input variables for a GP with Gaussian observation model.

A function for computing the KL-diff2 values for all observations
and input variables for a GP with Gaussian observation model.

A function for computing the KL-diff2 values for all observations
and input variables for a GP with Gaussian observation model when the
model of interest is a linear model. This version uses a linear kernel
in the GP in addition to the squared exponential kernel.
}
